requirements.txt
psycopg2-binary
bs4
requests